+++ editors/emacs/lambdapi-vars.el
@@ -15,6 +15,7 @@
     "reflexivity"
     "remove"
     "rewrite"
+    "set"
     "simplify"
     "solve"
     "symmetry"