+++ CHANGES.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Add tactic `set`.
+
+### Fixed
+
+- Induction tactic.
+
 ## 2.5.1 (2024-07-22)
 
 ### Added